Move train and preprocess scripts
